separate test app out of the runner
