+++ packages/runner/package.json
@@ -9,8 +9,7 @@
     "start": "runner",
     "prepare": "cd frontend && npm install",
     "lint": "eslint . --ignore-path .gitignore --ext .js",
-    "build": "cd frontend && npm run build",
-    "dev": "runner -Dde --dev"
+    "build": "cd frontend && npm run build"
   },
   "repository": {
     "type": "git",
@@ -30,8 +29,6 @@
   ],
   "description": "",
   "dependencies": {
-    "@urturn/test-app-backend": "*",
-    "@urturn/test-app-frontend": "*",
     "chalk": "^4.1.2",
     "commander": "^9.1.0",
     "cors": "^2.8.5",
