--- packages/test-app-backend/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@urturn/test-app-backend",
-  "version": "0.0.1",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC"
-}
