+++ packages/test-app/frontend/public/favicon.ico
