+++ packages/test-app/frontend/public/manifest.json
